+++ src/gui.js
@@ -86,7 +86,7 @@ BlockVisibilityDialogMorph, ThreadManager*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2022-April-22';
+modules.gui = '2022-April-25';
 
 // Declarations
 
@@ -2471,7 +2471,7 @@ IDE_Morph.prototype.endBulkDrop = function () {
     this.bulkDropInProgress = false;
 };
 
-IDE_Morph.prototype.droppedImage = function (aCanvas, name, embeddedCode) {
+IDE_Morph.prototype.droppedImage = function (aCanvas, name, embeddedData) {
     var costume = new Costume(
         aCanvas,
         this.currentSprite.newCostumeName(
@@ -2491,7 +2491,7 @@ IDE_Morph.prototype.droppedImage = function (aCanvas, name, embeddedCode) {
         return;
     }
 
-    costume.code = embeddedCode || null;
+    costume.embeddedData = embeddedData || null;
     this.currentSprite.addCostume(costume);
     this.currentSprite.wearCostume(costume);
     this.spriteBar.tabBar.tabTo('costumes');
@@ -9923,7 +9923,7 @@ CostumeIconMorph.prototype.createThumbnail = function () {
     var watermark, txt;
     SpriteIconMorph.prototype.createThumbnail.call(this);
     watermark = this.object instanceof SVG_Costume ? 'svg'
-        : (this.object.code ? '</>' : null);
+        : (this.object.embeddedData ? '</>' : null);
     if (watermark) {
         txt = new StringMorph(
             watermark,
@@ -9973,8 +9973,8 @@ CostumeIconMorph.prototype.userMenu = function () {
     menu.addItem("duplicate", "duplicateCostume");
     menu.addItem("delete", "removeCostume");
     menu.addLine();
-    if (this.object.code) {
-        menu.addItem("get blocks", "importCode");
+    if (this.object.embeddedData) {
+        menu.addItem("get blocks", "importEmbeddedData");
     }
     menu.addItem("export", "exportCostume");
     return menu;
@@ -10055,8 +10055,8 @@ CostumeIconMorph.prototype.removeCostume = function () {
     }
 };
 
-CostumeIconMorph.prototype.importCode = function () {
-    this.parentThatIsA(IDE_Morph).droppedText(this.object.code);
+CostumeIconMorph.prototype.importEmbeddedData = function () {
+    this.parentThatIsA(IDE_Morph).droppedText(this.object.embeddedData);
 };
 
 CostumeIconMorph.prototype.exportCostume = function () {
@@ -10064,8 +10064,8 @@ CostumeIconMorph.prototype.exportCostume = function () {
     if (this.object instanceof SVG_Costume) {
         // don't show SVG costumes in a new tab (shows text)
         ide.saveFileAs(this.object.contents.src, 'text/svg', this.object.name);
-    } else if (this.object.code) {
-        // embed blocks code inside the PNG image data
+    } else if (this.object.embeddedData) {
+        // embed payload data (e.g blocks)  inside the PNG image data
         ide.saveFileAs(this.object.pngData(), 'image/png', this.object.name);
     } else { // rasterized Costume
         ide.saveCanvasAs(this.object.contents, this.object.name);